Fixing a couple of warnings.

+++ fluid/shell_command.cxx
@@ -365,7 +365,7 @@ void shell_pipe_cb(FL_SOCKET, void*) {
 static void expand_macro(std::string &cmd, const std::string &macro, const std::string &content) {
   for (int i=0;;) {
     i = cmd.find(macro, i);
-    if (i==std::string::npos) break;
+    if (i==(int)std::string::npos) break;
     cmd.replace(i, macro.size(), content);
   }
 }
